requirements.txt
# For scripts/audio-to-lrc.py – generate LRC lyrics from audio via Whisper
openai-whisper>=20231117
